refactor: Move contexts from App to AppWrapper
