+++ src/components/App.jsx
@@ -10,8 +10,6 @@ import { useI18n, useBreakpoints } from 'cozy-ui/transpiled/react'
 import { Sprite as IconSprite } from 'cozy-ui/transpiled/react/Icon'
 import flag, { FlagSwitcher } from 'cozy-flags'
 
-import { SelectedGroupProvider } from './Contexts/SelectedGroup'
-import { SearchProvider } from './Contexts/Search'
 import withContactsMutations from '../connections/allContacts'
 import ContactsSelectionBar from './layout/ContactsSelectionBar'
 import { ModalManager } from '../helpers/modalManager'
@@ -31,27 +29,23 @@ const ContactsApp = props => {
   }, [cleanTrashedGroups])
 
   return (
-    <SelectedGroupProvider>
-      <SearchProvider>
-        <Layout monocolumn="true">
-          {isMobile && (
-            <BarCenter>
-              <Title>
-                <span className={'fil-path-title'}>Contacts</span>
-              </Title>
-            </BarCenter>
-          )}
-          <Main>
-            {flag('switcher') && <FlagSwitcher />}
-            <ContactsSelectionBar trashAction={deleteContact} />
-            <ContentWrapper />
-            <Alerter t={t} />
-            <ModalManager />
-          </Main>
-          <IconSprite />
-        </Layout>
-      </SearchProvider>
-    </SelectedGroupProvider>
+    <Layout monocolumn="true">
+      {isMobile && (
+        <BarCenter>
+          <Title>
+            <span className={'fil-path-title'}>Contacts</span>
+          </Title>
+        </BarCenter>
+      )}
+      <Main>
+        {flag('switcher') && <FlagSwitcher />}
+        <ContactsSelectionBar trashAction={deleteContact} />
+        <ContentWrapper />
+        <Alerter t={t} />
+        <ModalManager />
+      </Main>
+      <IconSprite />
+    </Layout>
   )
 }
