+++ src/targets/browser/setupApp.jsx
